EntityFrameworkCore FluentApi + Configuration + Examples DONE

+++ LactafarmaAPI/LactafarmaAPI/src/LactafarmaAPI.Data/Entities/Alert.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -8,9 +9,11 @@ namespace LactafarmaAPI.Data.Entities
 {
     public class Alert
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public DateTime Created { get; set; }
-        public long DrugId { get; set; }
+        public int DrugId { get; set; }
+
+        //Navigation Properties
         public Drug Drug { get; set; }
         public virtual ICollection<AlertMultilingual> AlertsMultilingual { get; set; }
