+++ LactafarmaAPI/LactafarmaAPI/src/LactafarmaAPI.Data/Entities/AlertMultilingual.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -8,11 +9,13 @@ namespace LactafarmaAPI.Data.Entities
 {
     public class AlertMultilingual
     {
-        public long AlertId { get; set; }
+        public int AlertId { get; set; }
         public string Name { get; set; }
         public string Risk { get; set; }
         public string Comment { get; set; }
         public Guid LanguageId { get; set; }
+
+        //Navigation Properties
         public Alert Alert { get; set; }
         public Language Language { get; set; }
     }
