+++ LactafarmaAPI/LactafarmaAPI/src/LactafarmaAPI.Data/Entities/Alias.cs
@@ -1,12 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 
 namespace LactafarmaAPI.Data.Entities
 {
     public class Alias
     {
-        public long Id { get; set; }
-        public long DrugId { get; set; }
+        public int Id { get; set; }
+        public int DrugId { get; set; }
+
+        //Navigation Properties
         public Drug Drug { get; set; }
         public virtual ICollection<AliasMultilingual> AliasMultilingual { get; set; }
