+++ LactafarmaAPI/LactafarmaAPI/src/LactafarmaAPI.Data/Entities/AliasMultilingual.cs
@@ -1,12 +1,15 @@
 ﻿using System;
+using System.ComponentModel.DataAnnotations.Schema;
 
 namespace LactafarmaAPI.Data.Entities
 {
     public class AliasMultilingual
     {
-        public long AliasId { get; set; }
+        public int AliasId { get; set; }
         public string Name { get; set; }
         public Guid LanguageId { get; set; }
+
+        //Navigation Properties
         public Alias Alias { get; set; }
         public Language Language { get; set; }
     }
