+++ LactafarmaAPI/LactafarmaAPI/src/LactafarmaAPI.Data/Entities/BrandMultilingual.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -7,10 +8,12 @@ namespace LactafarmaAPI.Data.Entities
 {
     public class BrandMultilingual
     {
-        public long BrandId { get; set; }
+        public int BrandId { get; set; }
         public string Name { get; set; }
         public Guid LanguageId { get; set; }
 
+
+        //Navigation Properties
         public Brand Brand { get; set; }
         public Language Language { get; set; }
