+++ LactafarmaAPI/LactafarmaAPI/src/LactafarmaAPI.Data/Entities/Favorite.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -9,8 +10,10 @@ namespace LactafarmaAPI.Data.Entities
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public long DrugId { get; set; }
+        public int DrugId { get; set; }
         public Guid UserId { get; set; }
+
+        //Navigation Properties
         public Drug Drug { get; set; }
         public User User { get; set; }
     }
