+++ LactafarmaAPI/LactafarmaAPI/src/LactafarmaAPI.Data/Entities/GroupMultilingual.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -7,9 +8,11 @@ namespace LactafarmaAPI.Data.Entities
 {
     public class GroupMultilingual
     {
-        public long GroupId { get; set; }
+        public int GroupId { get; set; }
         public Guid LanguageId { get; set; }
         public string Name { get; set; }
+
+        //Navigation Properties
         public Group Group { get; set; }
         public Language Language{ get; set; }
     }
